+++ openproject-export/CHANGELOG.md
@@ -17,3 +17,4 @@ See doc/COPYRIGHT.md for more details.
 # Changelog
 
 * `#<ticket number>` Create plugin
+* Added project attachments export feature